refactor: repository 도메인으로 이동
