+++ src/main/java/nextstep/subway/applicaion/StationQueryService.java
@@ -3,13 +3,12 @@ package nextstep.subway.applicaion;
 import lombok.RequiredArgsConstructor;
 import nextstep.subway.applicaion.dto.StationResponse;
 import nextstep.subway.domain.Station;
-import nextstep.subway.repository.StationRepository;
+import nextstep.subway.domain.StationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityNotFoundException;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 @Service
