+++ src/main/java/nextstep/subway/domain/StationRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.repository;
+package nextstep.subway.domain;
 
 import nextstep.subway.domain.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
